Switch from AWS STS to AWS Cognito Identity Pools.
This is the second big deviation from the upstream project (the first being dropping SAML support). While this change now requires the user to setup a Cognito Identity Pool, it still remains generic with regards to OIDC providers. The only difference is, that users now need to setup such a identity pool.