Init sign-in utility.
While the aws-cli-oidc tool is able to get us temporary credentials,
users might want to also gain access to the console as this is helpful
for debugging or protoyping with AWS resources. This is the task of
this new aws-sign-in tool.

The tool is purposely not integrated into the aws-cli-oidc tool as
it operates at an another layer, the layer of profiles. The
aws-cli-oidc tool does not know anything about profiles as it is
supposed to be part of a profile _definition_.

Yet, this tool is bundled here. This is because those two use-cases
(getting temporary CLI credentials and being able to login to the
console using them) are closely coupled such that it makes sense to
have both tools in a single distribution.