Replace graphic with ASCII.
While the png graphic was pretty and all, the ASCII art based graphic
has one big advatage: it makes the README self-contained. And this is
actually an issue as in the distributeion step (make dist) the README
is copied along the LICENSE and binary files. So we would need to
either copy the png file, too, or replace it with something that is
not depending on an external file. I decided for the latter option.