Cache AWS tokens per role.
Before the commit, the AWS token was saved independently of a role. In
that case, if a user requests temporary credentials initially with a
role A, the obtained AWS token will be cached (so far so good). However,
if the user wants to obtains the credentials for some other role B and
the cached AWS token is still valid, this cached token for role A will
be reused. With this commit, the AWS token are cached per role.